MachineListening UGens adapted for Windows
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@7088 a380766d-ff14-0410-b294-a243070f3f08

+++ Source/plugins/Loudness.cpp
@@ -213,7 +213,7 @@ void Loudness_dofft(Loudness *unit, uint32 ibufnum) {
 	//unit->m_phontotal= phontotal; 
 	//now to sones: 
 	/* from Praat: Excitation.c  Sones = 2 ** ((Phones - 40) / 10)  */
-	unit->m_sones= pow (2.0, (phontotal - 40) / 10);
+	unit->m_sones= pow ((float)2.0, (phontotal - 40) / 10);
 	
 	//printf("phontotal %f sones %f \n",phontotal, unit->m_sones);
